shadow-dom/elements-and-dom-objects/csshostrule-interface/csshostrule-attributes/test-002.html

<!DOCTYPE html>
<!--
Distributed under both the W3C Test Suite License [1] and the W3C
3-clause BSD License [2]. To contribute to a W3C Test Suite, see the
policies and contribution forms [3].

[1] http://www.w3.org/Consortium/Legal/2008/04-testsuite-license
[2] http://www.w3.org/Consortium/Legal/2008/03-bsd-license
[3] http://www.w3.org/2004/10/27-testcases
-->
<html>
<head>
<title>Shadow DOM Test: A_10_03_01_02</title>
<link rel="author" title="Sergey G. Grekhov" href="mailto:sgrekhov@unipro.ru">
<link rel="help" href="http://www.w3.org/TR/2013/WD-shadow-dom-20130514/#css-host-rule-attributes">
<meta name="assert" content="CSSHostRule Interface: cssRules of type CSSRuleList">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../../../testcommon.js"></script>
<link rel="stylesheet" href="/resources/testharness.css">
</head>
<body>
<div id="log"></div>
<script>
test(unit(function (ctx) {

	var d = newRenderedHTMLDocument(ctx);

    var host = d.createElement('div');
    d.body.appendChild(host);

	var s = createSR(host);

	var style = d.createElement('style');
    style.innerHTML = '' +
			'span { color: blue; }' +
			'@host {' +
			'div { color: red; }' +
			'}';
	d.head.appendChild(style);

	assert_equals(d.styleSheets[0].cssRules[1].type, 1001,	'Wrong cssRule type');

}), 'A_10_03_01_02_T01');
</script>
</body>
</html>
